:hammer:refactor: ProductCategoryList 엔티티에 price 필드 추가
